sync_video_s3/requirements.txt
ffmpeg-python>=0.2.0
watchdog>=2.1.6
python-dotenv>=0.19.0 